refactor: moved SometimesEquatable to its own file
